Create backToIE.bat
This script helps companies that still need to use internet explorer for some certain programs even in the latest’s versions of Windows 10 (22H2). Edge received an update that affected Internet Explorer, when the user clicks to open IE, it redirects automatically to Edge, doesn't letting the user access an older content, so this script fixes it.